Fix Codacy Issues - [SonarCSharp_S3216, SonarCSharp_S3898]
