+++ Zilon.Core/Zilon.Core/Commands/ICommand.cs
@@ -19,20 +19,4 @@
         /// </summary>
         void Execute();
     }
-
-    public struct CanExecuteCheckResult
-    {
-        public string FailureReason { get; set; }
-        public bool IsSuccess { get; set; }
-
-        public static CanExecuteCheckResult CreateFailed(string failureReason)
-        {
-            return new CanExecuteCheckResult { IsSuccess = false, FailureReason = failureReason };
-        }
-
-        public static CanExecuteCheckResult CreateSuccessful()
-        {
-            return new CanExecuteCheckResult { IsSuccess = true };
-        }
-    }
 }
